clean up api utils
