+++ dashboard/components/BoxPlot.jsx
@@ -3,8 +3,8 @@ import React, { useMemo } from "react";
 import { scaleBand, scaleLinear } from "@visx/scale";
 import { Group } from "@visx/group";
 import { Line, Bar } from "@visx/shape";
-import { Text, useText } from "@visx/text";
-import { Tooltip, useTooltip } from "@visx/tooltip";
+import { Text } from "@visx/text";
+import { useTooltip } from "@visx/tooltip";
 import { localPoint } from "@visx/event";
 import { useParentSize } from "@visx/responsive";
 import { AxisBottom, AxisLeft } from "@visx/axis";
